backend/requirements.txt
fastapi[standard]==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
motor==3.6.0
python-dateutil==2.9.0
passlib==1.7.4
python-jose[cryptography]==3.4.0
bcrypt==4.0.1